tests/wpt/tests/css/css-ui/parsing/caret-color-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 3: getComputedStyle().caretColor</title>
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#caret-color">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #target {
    color: blue;
  }
</style>
</head>
<body>
<div id="target"></div>
<script>
'use strict';
const currentColor = "rgb(0, 0, 255)";
test_computed_value("caret-color", "auto", currentColor);
test_computed_value("caret-color", "currentColor", currentColor);
test_computed_value("caret-color", "red", "rgb(255, 0, 0)");
</script>
</body>
</html>
